+++ lib/src/locale_manager.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'locale_params.dart';
 
 /// (en) A InheritedWidget for controlling locale within an app.
 /// /// Any widget that calls LocaleManager.of(context) from this
@@ -12,7 +13,7 @@ class LocaleManager extends InheritedWidget {
   final ValueNotifier<Locale> localeNotifier;
   final List<Locale> supportedLocales;
   final Locale? defLocale;
-  final Locale fallbackLocale;
+  final LocaleParams appLocale = LocaleParams();
 
   /// (en) Note: Do not call this constructor directly.
   /// This is meant to be called via LocalizedApp.
@@ -33,17 +34,17 @@ class LocaleManager extends InheritedWidget {
   /// defLocaleの設定がある場合はdefLocaleを最優先にした計算が実行されます。
   ///
   /// * [localeNotifier] : Notifier for detecting widget changes.
+  /// * [supportedLocales] : An array of app supported locales.
+  /// If no locale is available, the first locale in this set will be used
+  /// as a fallback locale.
   /// * [defLocale] : The default locale. If this locale is not null,
   /// and the device supports it, this locale will be forced,
   /// ignoring the device locale.
-  /// * [supportedLocales] : An array of app supported locales.
-  /// * [fallbackLocale] : The locale used when the app does not support
-  /// the device's locale.
-  const LocaleManager(
+  /// This locale must be an element of supportedLocales.
+  LocaleManager(
       {required this.localeNotifier,
       required this.supportedLocales,
       required this.defLocale,
-      required this.fallbackLocale,
       required super.child,
       super.key});
 
@@ -59,6 +60,7 @@ class LocaleManager extends InheritedWidget {
   /// (ja) このメソッドを経由してロケール情報にアクセスしたウィジェットは変更の追跡対象となり、
   /// ロケールの変更時に再ビルドされるようになります。
   static LocaleManager? of(BuildContext context) {
+    LocaleParams().context = context;
     return context.dependOnInheritedWidgetOfExactType<LocaleManager>();
   }
 
@@ -68,12 +70,15 @@ class LocaleManager extends InheritedWidget {
   /// fallbackLocale will be set.
   /// If defLocale is set, calculations will be performed with defLocale
   /// as the top priority.
+  /// The locale set by changeLocale will be cleared when this is called.
   ///
   /// (ja) ロケール設定を初期化します。
   /// 通常はデバイスロケールで初期化されますが、
   /// allowedLocalesにデバイスロケールが含まれない場合はfallbackLocaleが設定されます。
   /// defLocaleの設定がある場合はdefLocaleを最優先にした計算が実行されます。
+  /// changeLocaleで設定されたロケールは、これを呼び出すと解除されます。
   void reset() {
+    LocaleParams().useChangeLocale = false;
     if (defLocale != null) {
       if (supportedLocales.contains(defLocale!)) {
         changeLocale(defLocale!);
@@ -93,18 +98,23 @@ class LocaleManager extends InheritedWidget {
     if (supportedLocales.contains(deviceLocale)) {
       changeLocale(deviceLocale);
     } else {
-      changeLocale(fallbackLocale);
+      changeLocale(supportedLocales.first);
     }
   }
 
   /// (en) Change the app locale.
+  /// If you try to set an unsupported locale, nothing happens.
   ///
   /// (ja) アプリのロケールを変更します。
+  /// サポートされていないロケールを設定しようとした場合は何もおこりません。
   ///
   /// * [locale] : Changed locale.
   void changeLocale(Locale locale) {
+    LocaleParams().useChangeLocale = true;
     if (localeNotifier.value != locale) {
-      localeNotifier.value = locale;
+      if (supportedLocales.contains(locale)) {
+        localeNotifier.value = locale;
+      }
     }
   }
 
@@ -112,13 +122,46 @@ class LocaleManager extends InheritedWidget {
   ///
   /// (ja) アプリの現在のロケールを取得します。
   Locale getLocale() {
-    return localeNotifier.value;
+    if (LocaleParams().useChangeLocale || defLocale != null) {
+      return localeNotifier.value;
+    } else {
+      if (LocaleParams().context == null) {
+        return localeNotifier.value;
+      } else {
+        if (LocaleParams().context!.mounted) {
+          return Localizations.localeOf(LocaleParams().context!);
+        } else {
+          return localeNotifier.value;
+        }
+      }
+    }
+  }
+
+  /// (en) Get now app locale.
+  /// This method returns null until changeLocale is called.
+  /// In other words,
+  /// if you use it as the locale for a MaterialApp or CupertinoApp,
+  /// changes to the device's locale will be effective while the app is
+  /// running until changeLocale is called.
+  /// However, if defLocale is set, it takes precedence.
+  ///
+  /// (ja) アプリの現在のロケールを取得します。
+  /// このメソッドはchangeLocaleが呼び出されるまではnullを返します。
+  /// つまり、MaterialAppやCupertinoAppのロケールとして使用した場合、
+  /// changeLocale呼び出しまでアプリ起動中におけるデバイスのロケール変更が有効になります。
+  /// ただし、defLocaleが設定されている場合はdefLocaleが優先されます。
+  Locale? getLocaleForApp() {
+    if (LocaleParams().useChangeLocale || defLocale != null) {
+      return localeNotifier.value;
+    } else {
+      return null;
+    }
   }
 
   /// (en) Gets the current language code for the app.
   ///
   /// (ja) アプリの現在の言語コードを取得します。
   String getLanguageCode() {
-    return localeNotifier.value.languageCode;
+    return getLocale().languageCode;
   }
 }