+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0
+
+* Better compatibility with standard locale settings.
+* The fallbackLocale setting has been removed. Now the first supportedLocale acts as the fallbackLocale.
+* Added getLocaleForApp method to LocaleManager. This allows your app to accept changes to the device's locale while it is running.
+
 ## 1.0.1
 
 * Fixed README.